--- app/src/main/java/app/wristkey/DonateActivity.kt
@@ -1,99 +1,0 @@
-package app.wristkey
-
-
-import android.annotation.SuppressLint
-import android.content.Intent
-import android.net.Uri
-import android.os.Build
-import android.os.Bundle
-import android.view.View
-import android.widget.Button
-import android.widget.TextView
-import android.widget.Toast
-import androidx.annotation.RequiresApi
-import androidx.appcompat.app.AppCompatActivity
-import com.google.android.wearable.intent.RemoteIntent
-import wristkey.R
-import java.util.*
-
-class DonateActivity : AppCompatActivity() {
-
-    lateinit var mfaCodesTimer: Timer
-    lateinit var utilities: Utilities
-
-    private lateinit var clock: TextView
-    private lateinit var donate: Button
-
-    private lateinit var backButton: Button
-
-    @RequiresApi(Build.VERSION_CODES.M)
-    @SuppressLint("SetTextI18n")
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_donate)
-
-        utilities = Utilities(applicationContext)
-        mfaCodesTimer = Timer()
-        initializeUI()
-        startClock()
-
-    }
-
-    override fun onStop() {
-        super.onStop()
-        mfaCodesTimer.cancel()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        mfaCodesTimer.cancel()
-        finish()
-    }
-
-    override fun onStart() {
-        super.onStart()
-        mfaCodesTimer = Timer()
-    }
-
-    private fun startClock () {
-        if (!utilities.db.getBoolean(utilities.SETTINGS_CLOCK_ENABLED, true)) clock.visibility = View.GONE
-
-        try {
-            mfaCodesTimer.scheduleAtFixedRate(object : TimerTask() {
-                override fun run() {
-                    runOnUiThread { clock.text = utilities.getTime() }
-                }
-            }, 0, 1000)
-        } catch (_: IllegalStateException) { }
-    }
-
-    fun initializeUI () {
-        clock = findViewById(R.id.clock)
-
-        donate = findViewById(R.id.donate)
-        donate.setOnClickListener {
-            val uri: String = getString(R.string.donation_uri)
-            val intent = Intent(Intent.ACTION_VIEW).addCategory(Intent.CATEGORY_BROWSABLE).setData(Uri.parse(uri))
-            RemoteIntent.startRemoteActivity(this, intent, null)
-            Toast.makeText(this, "Opening GitHub donation page", Toast.LENGTH_SHORT).show()
-            try {
-                val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(uri))
-                startActivity(browserIntent)
-            } catch (_: Exception) { }
-        }
-
-        backButton = findViewById(R.id.backButton)
-        backButton.setOnClickListener {
-            finish()
-        }
-
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        when (requestCode) {
-            utilities.EXPORT_RESPONSE_CODE -> {  }
-        }
-    }
-
-}
