+++ app/src/main/res/layout/activity_donate.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@color/black"
     android:padding="@dimen/box_inset_layout_padding"
-    tools:context="app.wristkey.AddActivity"
+    tools:context="zeroxfourf.wristkey.AddActivity"
     tools:deviceIds="wear">
 
     <ScrollView
