+++ app/src/main/res/layout/activity_file_import.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:background="@color/black"
     android:padding="@dimen/box_inset_layout_padding"
-    tools:context="app.wristkey.FileImportActivity"
+    tools:context="zeroxfourf.wristkey.FileImportActivity"
     tools:deviceIds="wear">
 
     <ScrollView
