+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context="app.wristkey.MainActivity"
+    tools:context="zeroxfourf.wristkey.MainActivity"
     tools:deviceIds="wear">
 
     <androidx.core.widget.NestedScrollView
