+++ app/src/main/res/layout/activity_qrcode.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@color/black"
     android:padding="@dimen/box_inset_layout_padding"
-    tools:context="app.wristkey.AegisJSONImport"
+    tools:context="zeroxfourf.wristkey.AegisJSONImport"
     tools:deviceIds="wear">
 
     <ScrollView
